01.2 Part B Homework
