+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.example.hellotoast"
+        applicationId "com.example.helloconstraint"
         minSdkVersion 26
         targetSdkVersion 30
         versionCode 1
