+++ app/src/main/res/values/colors.xml
@@ -3,4 +3,6 @@
     <color name="colorPrimary">#6200EE</color>
     <color name="colorPrimaryDark">#3700B3</color>
     <color name="colorAccent">#03DAC5</color>
+    <color name="darkGrey">#FFAAAAAA</color>
+    <color name="yellow">#FFEB3B</color>
 </resources>
